requirements/base.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements/base.txt requirements/base.in
#
anyascii==0.3.0
    # via wagtail
asgiref==3.5.0
    # via django
beautifulsoup4==4.9.3
    # via
    #   django-htmlmin
    #   wagtail
certifi==2021.10.8
    # via
    #   elasticsearch
    #   requests
    #   sentry-sdk
charset-normalizer==2.0.10
    # via requests
deprecated==1.2.13
    # via redis
django==3.2.11
    # via
    #   django-filter
    #   django-redis
    #   django-taggit
    #   django-treebeard
    #   djangorestframework
    #   wagtail
django-filter==21.1
    # via
    #   -r requirements/base.in
    #   wagtail
django-htmlmin==0.11.0
    # via -r requirements/base.in
django-modelcluster==5.2
    # via wagtail
django-redis==5.2.0
    # via -r requirements/base.in
django-taggit==1.5.1
    # via wagtail
django-treebeard==4.5.1
    # via wagtail
django-treenode==0.17.0
    # via -r requirements/base.in
djangorestframework==3.13.1
    # via wagtail
draftjs-exporter==2.1.7
    # via wagtail
elasticsearch==7.16.1
    # via -r requirements/base.in
et-xmlfile==1.1.0
    # via openpyxl
gunicorn==20.0.4
    # via -r requirements/base.in
html5lib==1.1
    # via
    #   django-htmlmin
    #   wagtail
idna==3.3
    # via requests
l18n==2021.3
    # via wagtail
openpyxl==3.0.9
    # via tablib
packaging==21.3
    # via redis
pillow==9.0.0
    # via wagtail
psycopg2-binary==2.9.2
    # via -r requirements/base.in
pydantic==1.8.2
    # via -r requirements/base.in
pyparsing==3.0.7
    # via packaging
python-dateutil==2.8.2
    # via -r requirements/base.in
pytz==2021.3
    # via
    #   django
    #   django-modelcluster
    #   djangorestframework
    #   l18n
redis==4.1.2
    # via django-redis
requests==2.27.1
    # via
    #   wagtail
    #   wagtail-generic-chooser
sentry-sdk==1.5.1
    # via -r requirements/base.in
six==1.16.0
    # via
    #   django-treenode
    #   html5lib
    #   l18n
    #   python-dateutil
soupsieve==2.3.1
    # via beautifulsoup4
sqlparse==0.4.2
    # via django
tablib[xls,xlsx]==3.2.0
    # via wagtail
telepath==0.2
    # via wagtail
typing-extensions==4.0.1
    # via pydantic
typus==0.2.2
    # via wagtail-typograf
unidecode==1.3.2
    # via -r requirements/base.in
urllib3==1.26.7
    # via
    #   elasticsearch
    #   requests
    #   sentry-sdk
wagtail==2.15.2
    # via -r requirements/base.in
wagtail-generic-chooser==0.3.1
    # via wagtailsvg
wagtail-instance-selector==2.1.1
    # via -r requirements/base.in
wagtail-typograf==0.1.3
    # via -r requirements/base.in
wagtailsvg==0.0.28
    # via -r requirements/base.in
webencodings==0.5.1
    # via html5lib
willow==1.4
    # via wagtail
wrapt==1.13.3
    # via deprecated
xlrd==2.0.1
    # via tablib
xlsxwriter==3.0.2
    # via wagtail
xlwt==1.3.0
    # via tablib

# The following packages are considered to be unsafe in a requirements file:
# setuptools
